requirements.txt
# Your requirements go here
art-6.0